+++ CHANGES.rst
@@ -9,6 +9,8 @@ in progress
 - Improve HTTP response caching
 - Probe Redis before starting and croak if connection fails
 - Add "geohash" field when writing into InfluxDB
+- Use nominatim.hiveeyes.org as primary reverse geocoder,
+  fall back to nominatim.openstreetmap.org
 
 
 2018-12-11 0.4.3